+++ R/summaryFun.R
@@ -10,7 +10,9 @@
 #'
 #' @param data A data frame or tibble. Occurrence records to use as input.
 #' @param dontFilterThese A character vector of flag columns to be ignored in the creation or updating
-#'  of the .summary column.
+#'  of the .summary column. Cannot be specified with onlyFilterThese.
+#' @param onlyFilterThese A character vector. The inverse of dontFilterThese, where columns identified
+#' here will be filtered and no others. Cannot be specified with dontFilterThese.
 #' @param removeFilterColumns Logical. If TRUE all columns starting with "." will be removed in the 
 #' output data. This only makes sense to use when filterClean = TRUE. Default = FALSE.
 #' @param filterClean Logical. If TRUE, the data will be filtered to only those occurrence where .summary
@@ -50,6 +52,7 @@
 summaryFun <- function(
     data = NULL,
     dontFilterThese = NULL,
+    onlyFilterThese = NULL,
     removeFilterColumns = FALSE,
     filterClean = FALSE){
   # locally bind variables to the function
@@ -58,6 +61,19 @@ summaryFun <- function(
   #### 0.0 Prep ####
   if(is.null(data)){
     stop("You must provide a dataset in the 'data' argument.")
+  }
+  if(!is.null(dontFilterThese) & !is.null(onlyFilterThese)){
+    stop("Please only choose dontFilterThese OR onlyFilterThese.")
+  }
+  
+  ##### 0.1 onlyFilter to dontFilter ####
+    # In order to use onlyFilterThese, simply transform it to dontFilterThese
+  if(!is.null(onlyFilterThese)){
+    dontFilterThese <- data %>%
+      # Select all columns starting with "."
+      dplyr::select(tidyselect::starts_with(".")) %>%
+      colnames(.) %>%
+      setdiff(., onlyFilterThese)
   }