+++ tests/testthat/test-beesChecklist.R
@@ -6,7 +6,7 @@ system.file("extdata", "testChecklist.rda", package="BeeBDC") |>
 
 # Test the expected results
 testthat::test_that("testChecklist expected number of columns", {
-  testthat::expect_equal(ncol(testChecklist), 23)
+  testthat::expect_equal(ncol(testChecklist), 24)
 })
 
 # Test classes
@@ -14,7 +14,7 @@ testthat::test_that("testChecklist expected class", {
   testthat::expect_type(testChecklist, "list")
 })
 testthat::test_that("testChecklist expected class", {
-  testthat::expect_equal(attributes(testChecklist)$class, c("spec_tbl_df", "tbl_df","tbl","data.frame" ))
+  testthat::expect_equal(attributes(testChecklist)$class, c("tbl_df","tbl","data.frame" ))
 })
 
 
@@ -28,7 +28,7 @@ beesChecklist <- BeeBDC::beesChecklist()
 if(!is.null(beesChecklist)){
 # Test the expected results
 testthat::test_that("testChecklist expected number of columns", {
-  testthat::expect_equal(ncol(testChecklist), 23)
+  testthat::expect_equal(ncol(testChecklist), 24)
 })
 
 # Test classes
@@ -36,7 +36,7 @@ testthat::test_that("testChecklist expected class", {
   testthat::expect_type(testChecklist, "list")
 })
 testthat::test_that("testChecklist expected class", {
-  testthat::expect_equal(attributes(testChecklist)$class, c("spec_tbl_df", "tbl_df","tbl","data.frame" ))
+  testthat::expect_equal(attributes(testChecklist)$class, c("tbl_df","tbl","data.frame" ))
 })
 
 } #END !is.null